Add module to handle threading
This module provides a common interface for getting and setting the
number of threads in low-level threaded packages (OpenBLAS, MKL). We
also provide a function for disabling the behavior of these packages
where they use multiple threads if the user doesn't explicitly state
the number of desired threads, which increases contention when
parallelising at a higher level (e.g., python multiprocessing).