add test of package versions
Unfortunately, we're somewhat limited in what we can test because we
only get the versions of things being used at runtime, and this
package sits rather low in the dependency chain so there's not usually
a lot of other packages available when this test gets run.  Therefore
some of the tests are only checking that things don't explode, since
they are incapable of testing much more than that. But better that
than nothing.